Threading now implemented, with the number of threads being equal to Npara. Tested currently on macOS, needs testing on proper Linux, but attempts have been made to be compatible. Attempt to use CPU affinity to distribute threads on to different cores. However, on current tests with macOS gosia still doesn't run in parallel, even though they are correctly isolated in their own threads; it seems that we just wait for each thread to finish before starting the next one.
Some further streamlining of function calls to stop so many return parameters and checks. We now just exit if there is a failure.

By default, if an integration file is not given, the integration will not be performed, which is discouraged. It will no longer attempt to find the file and a warning will be given at execution if the filename is missing.